ui: background screen ui completed
