+++ lib/views/background_sound/components/volume_slider_component.dart
@@ -1,0 +1,38 @@
+import 'package:Medito/constants/constants.dart';
+import 'package:Medito/view_model/background_sounds/background_sounds_viewmodel.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'custom_track_shape_component.dart';
+
+class VolumeSliderComponent extends ConsumerWidget {
+  const VolumeSliderComponent({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final bgSoundNotifierProvider = ref.watch(backgroundSoundsNotifierProvider);
+    var currentVolume = bgSoundNotifierProvider.volume;
+    return SliderTheme(
+      data: SliderThemeData(
+          trackShape: CustomTrackShapeComponent(
+              leadingTitle: 'Volume',
+              tralingText: currentVolume.toString().split('.').first + '%'),
+          overlayShape: RoundSliderOverlayShape(overlayRadius: 0.0),
+          thumbShape: RoundSliderThumbShape(enabledThumbRadius: 0.0),
+          trackHeight: 72),
+      child: Slider(
+        value: currentVolume,
+        min: 0,
+        max: 100,
+        divisions: 100,
+        activeColor: ColorConstants.purple,
+        inactiveColor: ColorConstants.greyIsTheNewGrey,
+        onChanged: (double newValue) {
+          bgSoundNotifierProvider.handleOnChangeVolume(newValue);
+        },
+        semanticFormatterCallback: (double newValue) {
+          return '${newValue.round()} ';
+        },
+      ),
+    );
+  }
+}
